Rename command handler
