+++ lib/commands/index.js
@@ -3,7 +3,7 @@
 const middleware = require('../middleware');
 const subscribe = require('./subscribe');
 const signin = require('./signin');
-const changeState = require('./change-state');
+const issueState = require('./issue-state');
 const listSubscriptions = require('./list-subscriptions');
 const help = require('./help');
 const debug = require('./debug');
@@ -39,7 +39,7 @@ module.exports = (robot) => {
     middleware.authenticate,
     middleware.getResource(),
     middleware.getInstallation,
-    changeState.bind(null, 'closed'),
+    issueState.bind(null, 'closed'),
   );
 
   app.post(
@@ -48,7 +48,7 @@ module.exports = (robot) => {
     middleware.authenticate,
     middleware.getResource(),
     middleware.getInstallation,
-    changeState.bind(null, 'open'),
+    issueState.bind(null, 'open'),
   );
 
   app.post('/signin', signin);
